fix(auth): resolve type conflicts by consolidating auth type declarations
- Move type declarations from types.d.ts to auth.config.ts to avoid naming conflicts
- Use proper type aliasing (User as UserEntity) to prevent ambiguity
- Fix JWT module path to use @auth/core/jwt instead of next-auth/jwt
- Ensure teams property is properly handled in JWT and session callbacks
